refactor(cli): import the delivery route name default from core
`DEFAULT_DELIVERY_ROUTE_NAME` was declared twice: in the attachments
engine, and again in `packages/cli/src/attachments-check.ts` as an
independent const with the same name and value.

The default is not private to the engine, which is why the duplication
matters more than duplication usually does. `guren check`'s attachments
rules ask, from @guren/cli, whether the name the delivery route registers
under is claimed by more than one route -- `Router.name()` silently
overwrites duplicates, so a collision resolves route() lookups and typed
links to whichever registered last. To ask that question the rule has to
name the same route name the engine applies. Nothing made the two agree,
and the way they come apart is silent: if the framework's default moved,
the rule over there would not fail loudly, it would stop matching the
route that was actually registered, find one claim instead of two, and
report a genuine collision as fine.

(The neighbouring "is the delivery route mounted at all" rule reads the
controller class, not the name, so it was never exposed to this drift.)

The constant is now re-exported through `attachments/index.ts` and named
in core's barrel (an allowlist for these names, not `export *`) so the
rule can import the default rather than restate it.
`DEFAULT_DELIVERY_PREFIX` deliberately stays unexported: it is the other
value `resolveDeliveryRoute()` defaults, but nothing outside this package
names a delivery URL today, and a name in that allowlist is a semver
commitment -- it goes out when a rule over there needs it.

Two tests keep the single source structural: the constant must stay
reachable from `@guren/core`, and the engine's source may not re-hardcode
the name. The second is a source-level check on purpose -- nothing at
runtime distinguishes a name read from the constant from one re-typed as
a literal, the same reason the MCP endpoint gate pins the source form of
`process.env.NODE_ENV`.

That second test needs the same comment-blanking scanner #633 wrote for
the object-key prefix, so the scanner moves to `tests/source-scan.ts`
rather than being copied. Two copies of a scanner in one package is the
bug these tests exist to prevent, and the copies had already begun to
differ in their prose.

One further effect, verified by mutation: the CLI's existing
duplicate-name test builds its fixture from the literal
'attachments.show' against a `delivery: {}` config, so it agreed with the
local copy no matter what the framework did. With the import in place,
renaming core's default turns that test red.

No behaviour changes.
