+++ packages/core/src/attachments/index.ts
@@ -2,7 +2,7 @@ export { Attachable } from './Attachable.js'
 export type { AttachableRecordId, AttachableStatic } from './Attachable.js'
 export { configureAttachments } from './configure.js'
 export type { ConfigureAttachmentsOptions, ConfiguredAttachments } from './configure.js'
-export { ATTACHMENT_OBJECT_PREFIX } from './engine.js'
+export { ATTACHMENT_OBJECT_PREFIX, DEFAULT_DELIVERY_ROUTE_NAME } from './engine.js'
 export type {
   AttachOptions,
   AttachmentUrlOptions,
